Added new command node class SetPaddingCallback
SetPaddingCallback set the callback which returns the length of the
padding to be added to the message in record layer.

It can set the callback as a custom callback provided by the user or
use one of the builtin callbacks provided in static methods.

A callback factory to which returns callbacks for fixed length paddings
is provided.