Added new command node class AddRecordPadding
AddRecordPadding set the callback which returns the length of the
padding to be added to the message in record layer.

It can set the callback as a custom callback provided by the user or
use one of the builtin callbacks.

If no custom callback is provided, but the size of the desired padding
is provided, the callback is set to a simple method which returns the
requested padding size.

If no custom callback is provided, but 'fill' is set to 'True', the
callback is set to a method which returns the maximum padding length
possible.

If nothing is provided, the callback is set to a method which returns
the value 16.