Add tools for collecting timing information
This commit allows tlsfuzzer to collect timing information and consists
of three main parts. First is the Log class, that just logs classes of the
test conversations to a file and can later read it. Second is the TimingRunner
class that is intentioned to run after the regular tests have passed and is
meant to re-run them multiple times with tcpdump running in the background.
Last is the Analysis module that at this time associates classes of tests
with the latency of last server response and can dump the data into a csv.
Later the Analysis module should also contain tools for statistical analysis.
This commit adds one optional dependency - dpkt (used to decode pcap files).

References #47