Add statistical analysis, expand test coverage, change CI conf.
This commit introduces statistical analysis along with plot
generation. Mocking of the filesystem was fixed and unit tests
were written to get above the requested 90%. Analysis was renamed
as Extract because it made more sense to split the responsibilites
into two modules. Extract now handles extracting timing info from
the pcap and Analysis handles statistical tests and plots.
CI config was changed to allow for testing with timing dependencies.
Changes are reflected in the docs.