tests/scripts_retention.py: Use sys.executable instead of fixed 'python'
This works regardless of the name of the installed interpreter, and
has the additional advantage of always using the same interpreter as
used to call the script.

The JSON test configurations now use a "{python}" token instead of
"python" to avoid confusion about it being replaced with
sys.executable at runtime.