switch to Wilcoxon signed-rank test
because we collect observations in sets and the collection happens
over long periods of time (so noise that is constant for
a significant portion of collection will be visible in a portion of
observations), it means that our samples *are* dependent

because KS-test requires independent samples, we can't use it and
expect correct results (in practice, KS-test overestimates p-values
for dependent samples). Wilcoxon signed-rank test does work with
dependent samples. The downside is that it won't see change in variance
between samples. The upside is that it is significantly more sensitive
to changes in expected value.