analysis: write the general statistics to file, report CI
Write the result of the test for uniformity of p-values to a text file

Add also the calculation (using bootstrapping) of the confidence interval
for the truncated mean of differences between the worst pair of samples
(ones that have the smallest p-value from Wilcoxon test).
Write that result to the text file too.