Add a TLSv1.3 test that ensures a zero content type is handled correctly.
In TLSv1.3, once record protection has been engaged, all records are sent
as application data records on the wire, with the real content type being
added to the end of the plaintext record, followed by optional zero-byte
padding. If the content type is zero (the plaintext consists only of bytes
with value zero) then an unexpected message alert must be sent.

Updates #355