scripts_retention.py: don't override PYTHONPATH if already set
To use scripts_retention.py in TLS library test suite, tlslite-ng
needs to be visible to the script; typically this can be done either
by using pip or creating a symlink inside tlsfuzzer submodule:
https://gitlab.com/gnutls/gnutls/-/blob/3.7.2/tests/suite/tls-fuzzer/tls-fuzzer-common.sh#L43

While the latter is preferred in CI environment (because it allows a
specific tlslite-ng revision to be pinned), it's still not ideal to
modify the tlsfuzzer submodule checkout.  This patch adds the third
option: the user can adjust PYTHONPATH before running the script,
which was previously not possible because of PYTHONPATH override in
the script itself.

Signed-off-by: Daiki Ueno <dueno@redhat.com>