tlsfuzzer/analysis.py: fix issue with workers
Bit size analysis fix:
Some times the slice that the worker gets might have a chunk of data
that will not have any entry of max_k_size in it. Then we are trying to
find one and eventually throw IndexError. Fixed that by checking that i
is in range.